+++ src/backend/apps/previewgen/__init__.py
@@ -1,0 +1,5 @@
+from .previewgen import ( # NOQA401
+    create_extrusion_component_mesh,
+    convert_mesh_component_mesh,
+    create_component_preview
+)