+++ CHANGELOG.md
@@ -117,3 +117,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix linting in all components.
 - Change Sidebar width to `250px`
+
+## [0.0.1.6] - 2024-06-06
+
+### Versions
+
+- CSC FastAPI Backend:  0.1.0.3
+- CSC React Frontend:   0.1.4.3
+- CSC Sheetscan Module: 0.0.1.10
+
+### Added
+
+#### CSC FastAPI Backend
+
+- Added preview generation routine
+
+### Changed
+
+#### CCSC FastAPI Backend
+
+- Moved database connection functions to `utility` module